Callback and Import Fix
Added a callback custom transformation function so that the user does not
need to use the same code to transform the training dataset and the
prediction dataset in their pipeline before feeding them to PreSci.

Now just declare the transformation function with the dataset as
a parameter, transform what you want, return the dataset in the
end of the function, and pass that function to the "callback"
parameter of PreSci.

Fixed imports, that were scrambled because this project started inside
another machine learning project.