Import Hell
I found that to make the imports work as a submodule, they stopped working
as a standalone.

So the solution is to put all the code inside another package, this will
probably make importing as a submodule a little verbose
(from presci.app.presci import PreSci) but it will work in all intended ways.