Added Custom Encoder!
- Added custom encoder, in which you can pass a dictionary with first
level keys as variable names, second level keys as current variable
values and second level values as values to encode the current ones.

- Added exception if the user passes the same variable to do more than
one form of encoding.

- Fixed bug where plotted categorical features would not show Rare values
as "Rare", and instead would show them with their original label. This is
a problem because, if the user is going to custom encode the labels, he
must know the relations of the "Rare" values, (if there are any) so that
he can encode them as well, otherwise those values labeled as "Rare" will
be treated as unknown values by the encoder, which means setting them
as "-1".

- Fixed a bug where sometimes the predicted data would be completely wrong
because sklearn's experimental IterativeImputer depends on the order of the
columns, so I'm sorting the column order before feeding data to the MICE
model for training, and for prediction.