Fixed MICE Malfunction Elegantly
 The MICE model, which is currently sklearn's implementation of
the experimental IterativeImputer is dependant on the order of columns
in the dataframe.

 The problem is that, in the "transform" method (the one that transforms
for prediction), I select only the features, this is something I did thinking
of a production environment, in case someone inputs a dataframe with more
columns than are required, so that the model doesn't break.

 This selection is done by a list stored in the meta_data dictionary by
the Analyzer module, the thing is that this list was being stored in a
random order (the order of the dictionary), and so when PreSci selected
only the features (using DataFrame.loc) it was also changing the order of
the columns.

 Now the columns will be always ordered because the list in the meta_data
is ordered, since I'm no longer storing it in the dictionary's order, and
actually using the training DataFrame's order. Hoorray!