Complete Overhaul of PreSci's API and Scope
I started this project believing there might be an specific order
and way to transform features according to some criteria. I now realize
that this is a much deeper subject, and also very subjective, so PreSci
has to adapt and become much more flexible. It is with this mindset that
this rewriting has taken place.

The idea now is that PreSci will be modular, users can use the Analyzer
to get meta-data about a dataset, and to plot it's features, and use the
Transformer to replace Rare labels, exclude outliers, to fit encoder, MICE,
embedder, and scaler models to data (and of course, use them to encode,
replace missing, and scale). This results in a more flexible tool, which
I'm hoping to be useful in a broader spectrum of projects.

Some of the functionalities require only one method call, while some require
two (the ones where a model must be fitted or the Transformer object
must save some data).

- Methods that execute the entire functionality by themselves are:

    remove_outliers()
    log()
    split()

- Methods whose functionality will be partialy executed are:

    set_frequent_labels(), to be followed by: replace_rare_labels()

    fit_onehot_encoder(), to be followed by: onehot_encode()

    fit_ordinal_encoder(), to be followed by: ordinal_encode()

    fit_auto_encoder_continuous_target() or fit_auto_encoder_boolean_target(),
    to be followed by: auto_encode()

    set_custom_encoder(), to be followed by: custom_encode()

    fit_mice(), to be followed by: replace_missing()

    fit_embedder(), to be followed by: embed()

    fit_minmax_scaler() and fit_standard_scaler(),
    to be followed by: scale()

Basically the API of those functions follows the following rules:

    1- All of the fit and set functions receive a DataFrame and fit/set for
    all variables in the DataFrame

    2- All of the functions that follow them receive a DataFrame and only
    change variables inputed in the previous function

This means that you can use the following functions with the entire dataset,
or with only the fitted columns, check the pydocs in the module for further
instructions.

- Updated and expanded the existing documentation, including examples
for public methods

- Reorganized code so that public methods have a separate section on top,
so that it is more intuitive to read

- MICE functionality now saves the order of columns in the DataFrame
where it was fitted, so that the columns can enter in any order for
replacing missing data.

- Added Exception triggered by parameter type checking

- Fixed embedding Neural Network